pyproject.toml
[project]
name = "seamless-s2st-ws"
version = "0.1.0"
description = "FastAPI WebSocket speech-to-speech translation (CPU or ROCm)"
authors = [{ name = "Jesper", email = "jespertk@gmail.com" }]
requires-python = ">=3.13,<3.14"
readme = "README.md"
dependencies = [
    "audioop-lts>=0.2.2",
    "fastapi>=0.116.1",
    "uvicorn[standard]>=0.35",
    "pydantic>=2.11.7",
    "numpy>=2.3.2,<2.4.0",
    "transformers>=4.55.2,<4.58",
    "webrtcvad>=2.0.10",
    "silero>=0.4.1",
    "soundfile>=0.13.1",
    "g711>=1.0.0",
    "httpx>=0.28.1",
    "protobuf>=6.31.1",
    "sentencepiece>=0.2.1",
    "tiktoken>=0.11.0",
    "azure-cognitiveservices-speech==1.41.1",
]

[project.optional-dependencies]
# Plain CPU wheels from PyPI. Keep torch/torchaudio versions matched.
cpu = [
    "torch==2.7.0",
    "torchaudio==2.7.0",
]
# ROCm wheels + Triton, pinned to matching versions.
gpu-rocm = [
    "torch==2.7.0+rocm6.3", # Do NOT update to 2.8.0 as it will crash
    "torchaudio==2.7.0+rocm6.3",
    "pytorch-triton-rocm>=3.0.0",
]

[dependency-groups]
dev = [
    "mypy>=1.17.1,<2.0.0",
    "ruff>=0.12.8,<1",
]

[tool.uv]
# Ensures editable install behavior and deterministic resolution.
index-strategy = "unsafe-best-match"
extra-index-url = [
    "https://download.pytorch.org/whl/rocm6.3",
]

[tool.ruff]
line-length = 120

[tool.pyright]
pythonVersion = "3.13"

[build-system]
requires = ["setuptools>=80"]
build-backend = "setuptools.build_meta"
